backend/requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
feedparser
python-dotenv
readability-lxml
lxml
lxml_html_clean
sqlalchemy
psycopg2-binary
alembic
gTTS
google-genai
google-auth
google-auth-oauthlib
razorpay
python-jose[cryptography]
bleach
sentry-sdk
prometheus-client
